pyproject.toml
[project]
name = "backoffice"
version = "0.1.0"

[tool.poetry]
packages = [{ include = "backoffice", from = "src" }]

[tool.poetry.dependencies]
python = "3.13.*"
fastapi = ">=0.118,<0.120"
uvicorn = { version = ">=0.30,<0.32", extras = ["standard"] }
pydantic = ">=2,<3"
sqlalchemy = ">=2.0,<3.0"
alembic = ">=1.12,<2.0"
asyncpg = ">=0.30,<0.32"
httpx = ">=0.25,<1.0"
python-dotenv = ">=1,<2"
python-jose = { version = ">=3.3,<4.0", extras = ["cryptography"] }
passlib = { version = ">=1.7,<2.0", extras = ["bcrypt"] }
python-multipart = ">=0.0.6,<1.0"
authlib = ">=1.2,<2.0"
boto3 = ">=1.34,<2.0"
pillow = ">=10.0,<11.0"
requests = ">=2.31,<3.0"
redis = ">=5,<6"
elasticsearch = ">=8.11,<9.0"
python-json-logger = ">=2.0,<3.0"

[tool.poetry.group.dev.dependencies]
black = ">=24,<26"
pytest = ">=7.4,<9"
pytest-asyncio = ">=0.21,<1.0"
pytest-cov = ">=4.1,<6.0"
isort = ">=5.12,<7"
pytest-html = ">=3.2,<4.0"
pytest-xdist = ">=3.3,<4.0"
pytest-mock = ">=3.11,<4.0"
pytest-timeout = ">=2.1,<3.0"
pytest-benchmark = ">=4.0,<5.0"
factory-boy = ">=3.3,<4.0"
faker = ">=19,<21"
responses = ">=0.23,<1.0"
freezegun = ">=1.2,<2.0"
aiosqlite = ">=0.19,<1.0"
aiokafka = ">=0.10,<1.0"
minio = ">=7.1,<8.0"
psutil = ">=5.9,<6.0"
memory-profiler = ">=0.61,<1.0"
bandit = ">=1.7,<2.0"
safety = ">=2.3,<3.0"
flake8 = ">=6,<7"
mypy = ">=1.7,<2.0"
pylint = ">=3,<4"
rich = ">=13,<14"
click = ">=8.1,<9.0"

[build-system]
requires = ["poetry-core>=1.9.0"]
build-backend = "poetry.core.masonry.api"